Create Soccer Player Analysis - Robert Lewandowski
So this project maps the Bayern Munich forward's performance in the years 2019/2020  (as on 5/22/202) to figure out his consistency. The program uses the following libraries: BeautifulSoup4, requests, Pandas, matplotlib and numpy. The process starts when we extract data from a sports website and store it into a CSV file, and read the file to plot a clean graph showing lewandowsky's performance. In my opinion he is much more consistent than his competitors in Germany, and he has got a bright career ahead as long as he keeps his form and doesn't get injured frequently.